refactor: Update sCoin constants and types
